// src/app/shared/generated/model/aIKnowledgeBase.ts
/**
 * onecx-ai-management-bff
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AIContext } from './aIContext';


export interface AIKnowledgeBase {
    modificationCount?: number;
    id: string;
    name?: string;
    description?: string;
    appId?: string;
    aiContext?: Array<AIContext>;
}